requirements.txt
colander==1.0
Django==1.8.3
django-admin-sortable==1.8.3
django-appconf==1.0.1
django-ckeditor==4.4.8
django-extensions==1.5.3
django-filer==1.1.1
django-filter==0.10.0
django-image-cropping==1.0.1
django-mandrill==0.1.0
django-mptt==0.7.4
django-paintstore==0.2
django-pipeline==1.5.3
django-polymorphic==0.7.1
django-suit==0.2.14
django-taggit==0.14.0
djangorestframework==3.1.1
docopt==0.4.0
easy-thumbnails==2.2
ecdsa==0.13
Fabric==1.10.1
futures==3.0.3
ipython==3.2.1
iso8601==0.1.10
mandrill==1.0.57
MySQL-python==1.2.5
paramiko==1.15.2
Pillow==2.8.1
pudb==2015.3
Py-Authorize==1.2.3.0
pycrypto==2.6.1
Pygments==2.0.2
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
translationstring==1.3
Unidecode==0.4.18
urwid==1.3.0
